RoutingをRESTfulに沿うように変更した
