+++ .rubocop.yml
@@ -90,6 +90,7 @@ Metrics/BlockLength:
     - "spec/**/*"
     - "lib/tasks/auto_annotate_models.rake"
     - "Guardfile"
+    - "config/routes.rb"
 
 # メソッド名とインスタンス変数の名前一致させない(予約語被りで外した経緯)
 Naming/MemoizedInstanceVariableName:
