+++ app/controllers/v1/temp_user/sign_up_controller.rb
@@ -1,0 +1,16 @@
+module V1
+  module TempUser
+    class SignUpController < ApplicationController
+      def create
+        temp_user = User.create!(temp_sign_up_params)
+        render json: temp_user, serializer: MeSerializer, status: :created
+      end
+
+      private
+
+        def temp_sign_up_params
+          params.require(:temp_user).permit(:birth_year, :area_id)
+        end
+    end
+  end
+end
